requirements.txt
paramiko==3.4.0
pymysql==1.1.0
kafka-python==2.0.2
requests==2.31.0
gitpython==3.1.43